Token cannot pass to websockify
Fix the issue: https://github.com/kanaka/websockify/issues/213.
When the websockify use --token-plugin, the spice_auto.html cannot pass the token value to websockify. Because websockfiy use the path, but the path variable is useless in spice_auto.html. So i added the token to the uri if token is not empty.